Spawn editing book photos in BookInfoActivity instead of AddEditBook
