+++ app/src/main/res/layout/content_book_info.xml
@@ -18,6 +18,19 @@
             android:orientation="vertical"
             android:padding="25dp">
 
+            <LinearLayout
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:gravity="center"
+                android:orientation="vertical">
+
+                <ImageView
+                    android:id="@+id/bookImage"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    app:srcCompat="@drawable/ic_book_black_150dp" />
+            </LinearLayout>
+
             <TextView
                 android:id="@+id/Title"
                 android:layout_width="wrap_content"
